refactor: move requires around and modularize DetailStatString
