+++ js/index.js
@@ -14,9 +14,6 @@ const Handlebars = require('handlebars');
 const fs = require('fs-extra');
 const cp = require('child_process');
 
-const moment = require('moment');
-const FormData = require('form-data');
-
 const HeroesDB = require('./js/database.js');
 const Parser = require('./parser/parser.js');
 const HeroesTalents = require('./js/heroes-talents.js');
@@ -42,81 +39,7 @@ const DetailStatList = require('./js/game-data/detail-stat-list');
 
 const PerMapStatList = require('./js/game-data/map-stats');
 
-const DetailStatString = {
-  'Takedowns' : 'Takedowns',
-  'KillParticipation' : 'Kill Participation', // special case
-  'Kills' : 'Kills',  // special case
-  'HighestKillStreak' : 'Highest Kill Streak',
-  'SoloKill' : 'Kills',
-  'VengeancesPerformed' : 'Vengeances',
-  'Assists' : 'Assists',
-  'Deaths' : 'Deaths',
-  'TimeSpentDead' : 'Time Dead',
-  'OutnumberedDeaths' : 'Deaths While Outnumbered',
-  'EscapesPerformed' : 'Escapes',
-  'HeroDamage' : 'Hero Damage',
-  'SiegeDamage' : 'Siege Damage',
-  'StructureDamage' : 'Structure Damage',
-  'MinionDamage' : 'Minion Damage',
-  'SummonDamage' : 'Summon Damage',
-  'CreepDamage' : 'Creep Damage',
-  'Healing' : 'Healing',
-  'SelfHealing' : 'Self Healing',
-  'ProtectionGivenToAllies' : 'Allied Shields',
-  'ClutchHealsPerformed' : 'Clutch Heals',
-  'DamageTaken' : 'Damage Taken',
-  'TeamfightDamageTaken' : 'Team Fight Damage Taken',
-  'TeamfightEscapesPerformed' : 'Team Fight Escapes',
-  'TeamfightHealingDone' : 'Team Fight Healing',
-  'TeamfightHeroDamage' : 'Team Fight Hero Damage',
-  'TimeCCdEnemyHeroes' : 'CC Time',
-  'TimeRootingEnemyHeroes' : 'Root Time',
-  'TimeSilencingEnemyHeroes' : 'Silence Time',
-  'TimeStunningEnemyHeroes' : 'Stun Time',
-  'TimeOnPoint' : 'Time on Point',
-  'OnFireTimeOnFire' : 'Time on Fire',
-  'ExperienceContribution' : 'XP Contribution',
-  'MercCampCaptures' : 'Merc Camp Captures',
-  'TownKills' : 'Town Kills',
-  'WatchTowerCaptures' : 'Watch Tower Captures',
-  'Role' : 'Role',
-  'AltarDamageDone' : 'Altar Damage Done',
-  'DamageDoneToImmortal' : 'Damage to Immortal',
-  'DragonNumberOfDragonCaptures' : 'Dragon Knights Captured',
-  'DragonShrinesCaptured' : 'Shrines Captured',
-  'BlackheartDoubloonsCollected' : 'Dubloons Held At End',
-  'BlackheartDoubloonsTurnedIn' : 'Dubloons Turned In',
-  'MinesSkullsCollected' : 'Skulls Collected',
-  'DamageDoneToShrineMinions' : 'Shrine Minion Damage',
-  'GardensPlantDamage' : 'Plant Damage',
-  'GardensSeedsCollected' : 'Seeds Collected',
-  'GemsTurnedIn' : 'Gems Turned In',
-  'NukeDamageDone' : 'Nuke Damage',
-  'CurseDamageDone' : 'Curse Damage',
-  'TimeInTemple' : 'Time On Temple',
-  'DamageDoneToZerg' : 'Damage Done to Zerg',
-  'KDA' : 'KDA',
-  'timeDeadPct' : 'Time Dead %',
-  'PPK' : 'People Per Kill',
-  'mercCaptures' : 'Mercenary Captures',
-  'mercUptime' : 'Mercenary Uptime',
-  'mercUptimePercent' : 'Mercenary Uptime %',
-  'timeTo10' : 'Time to Level 10',
-  'timeTo20' : 'Time to Level 20',
-  'avgTimeSpentDead' : 'Average Time Spent Dead',
-  'T1' : 'Time at Level 1',
-  'T2' : 'Time at Level 4',
-  'T3' : 'Time at Level 7',
-  'T4' : 'Time at Level 10',
-  'T5' : 'Time at Level 13',
-  'T6' : 'Time at Level 16',
-  'damageDonePerDeath' : 'Damage per Death',
-  'damageTakenPerDeath' : 'Damage Taken per Death',
-  'healingDonePerDeath' : 'Healing per Death',
-  'DPM' : 'DPM',
-  'HPM' : 'HPM',
-  'XPM' : 'XPM'
-};
+const DetailStatString = require('./js/game-data/detail-stat-string');
 
 const RoleColor = {
   'Assassin' : '#db2828',
