+++ js/player.js
@@ -1,3 +1,5 @@
+const moment = require('moment');
+
 var playerDetailID;
 var playerDetailStats;
 var playerDetailInfo;
@@ -317,7 +319,7 @@ function initPlayerPage() {
   let playerWidget = $(getTemplate('filter', '#filter-popup-widget-template').find('.filter-popup-widget')[0].outerHTML);
   playerWidget.attr('widget-name', 'player-filter');
   playerWidget.find('.filter-widget-team').addClass('is-hidden');
-  
+
   $('#filter-widget').append(playerWidget);
   initPopup(playerWidget);
 
@@ -361,7 +363,7 @@ function initPlayerPage() {
     type: 'line',
     data: {
       datasets: [{
-        
+
       }]
     },
     options: progressionStatGraphSettings
@@ -548,7 +550,7 @@ function processPlayerData(err, docs) {
 // callback for hero select menu
 function showHeroDetails(value, text, $selectedItem) {
   playerHeroMatchThreshold = parseInt($('#player-hero-thresh input').val());
-  
+
   if (value === 'all' || value === '') {
     DB.getHeroDataForPlayerWithFilter(playerDetailInfo._id, playerDetailFilter, function(err, docs) {
       playerDetailStats = summarizeHeroData(docs);
@@ -719,7 +721,7 @@ function renderHeroTalentsTo(hero, container, docs) {
     for (let i = 0; i < 7; i++) {
       // this should theoretically be in order
       let context = {};
-  
+
       if (i < keys.length) {
         context.img = Heroes.talentIcon(build.talents[keys[i]]);
         context.description = Heroes.talentDesc(build.talents[keys[i]]);
@@ -789,7 +791,7 @@ function renderPlayerSummary() {
 
     if (context.name === "")
       context.name = "Default";
-    
+
     context.games = playerDetailStats.skins[s].games;
     context.winPercent = playerDetailStats.skins[s].wins / context.games;
     context.formatWinPercent = formatStat('pct', context.winPercent);
@@ -964,7 +966,7 @@ function renderProgression() {
 
   if (axis1Stats[0] === '')
     axis1Stats = [];
-  
+
   if (axis2Stats[0] === '')
     axis2Stats = [];
 
@@ -973,7 +975,7 @@ function renderProgression() {
   // get the data
   let query = Object.assign({}, playerDetailFilter);
   query.ToonHandle = playerDetailInfo._id;
-  
+
   let value = $('#player-hero-select-menu').dropdown('get value');
   if (value !== 'all' && value !== '') {
     query.hero = value;
@@ -1046,12 +1048,12 @@ function renderProgression() {
           return -1;
         else if (parseInt(ad[0]) > parseInt(bd[0]))
           return 1;
-        
+
         if (parseInt(ad[1]) < parseInt(bd[1]))
           return -1;
         else if (parseInt(ad[1]) > parseInt(bd[1]))
           return 1;
-        
+
         return 0;
       });
     }
@@ -1063,7 +1065,7 @@ function renderProgression() {
           return 1;
 
         return 0;
-      })  
+      })
     }
 
     // update data arrays
@@ -1170,7 +1172,7 @@ function hashInterval(date, mode) {
   else if (mode === IntervalMode.Week) {
     let mdate = moment(date)
     let ident = date.getFullYear() + '-' + mdate.week();
-    
+
     return [ident, mdate];
   }
   else if (mode === IntervalMode.Season) {
@@ -1189,7 +1191,7 @@ function hashInterval(date, mode) {
     let versionString = date.m_major + '.' + date.m_minor + '.' + date.m_revision + ' (build ' + date.m_build + ')';
     return [versionString, date.m_build];
   }
-  
+
   //listen if you call this with an invalid mode i hope it crashes
   return [null, null];
 }
@@ -1518,12 +1520,12 @@ function printPlayerSummary(filename, sections) {
 function exportPlayerHeroCSV(file) {
   let query = Object.assign({}, playerDetailFilter);
   query.ToonHandle = playerDetailInfo._id;
-  
+
   let value = $('#player-hero-select-menu').dropdown('get value');
   if (value !== 'all' && value !== '') {
     query.hero = value;
   }
-  
+
   DB.getHeroData(query, function(err, docs) {
     exportHeroDataAsCSV(docs, file);
   });
